microservices/sla_breach_detector/requirements.txt
flask
requests
kafka-python
apscheduler
pyyaml
prometheus_client